requirements.txt
beautifulsoup4==4.6.0
discord.py==0.16.11
praw==4.5.1
wikipedia==1.4.0
aiohttp
websocket
chardet
googlemaps==2.5.1